+++ src/complex/big.rs
@@ -3943,7 +3943,7 @@ pub struct BorrowComplex<'a> {
 
 impl BorrowComplex<'_> {
     // unsafe because the lifetime is obtained from return type
-    unsafe fn from_raw<'a>(raw: mpc_t) -> BorrowComplex<'a> {
+    pub(crate) unsafe fn from_raw<'a>(raw: mpc_t) -> BorrowComplex<'a> {
         BorrowComplex {
             inner: ManuallyDrop::new(Complex::from_raw(raw)),
             phantom: PhantomData,